# Make filestream benchmark more precise and more re-usable