* Now the initialization time is not included in the actual benchmark timer
* The filestream runner is more re-usable now and can be used later for regular tests not just benchmarks

<!-- Type of change
Please label this PR with one of the following labels, depending on the scope of your change:
- Bug
- Enhancement
- Breaking change
- Deprecation
- Cleanup
- Docs
-->

## Checklist

<!-- Mandatory
Add a checklist of things that are required to be reviewed in order to have the PR approved

List here all the items you have verified BEFORE sending this PR. Please DO NOT remove any item, striking through those that do not apply. (Just in case, strikethrough uses two tildes. ~~Scratch this.~~)
-->

- [x] My code follows the style guidelines of this project
- [x] I have commented my code, particularly in hard-to-understand areas
~~- [ ] I have made corresponding changes to the documentation~~
~~- [ ] I have made corresponding change to the default configuration files~~
- [x] I have added tests that prove my fix is effective or that my feature works
~~- [ ] I have added an entry in `CHANGELOG.next.asciidoc` or `CHANGELOG-developer.next.asciidoc`.~~

## How to test this PR locally

```sh
cd ./filebeat/input/filestream
go test -run=none -bench='BenchmarkFilestream/filestream_.*_throughput.*' -benchmem -benchtime=100x
```

Before

```
BenchmarkFilestream/filestream_default_throughput-10
100          10454563 ns/op        13576216 B/op     181225 allocs/op
BenchmarkFilestream/filestream_fingerprint_throughput-10
100          13427490 ns/op        13746259 B/op     191416 allocs/op
```

After

```
BenchmarkFilestream/filestream_default_throughput-10
100          10587500 ns/op        13540146 B/op     180370 allocs/op
BenchmarkFilestream/filestream_fingerprint_throughput-10
100          11463956 ns/op        13700589 B/op     190376 allocs/op
```

<!-- Recommended
Explain here how this PR will be tested by the reviewer: commands, dependencies, steps, etc.
-->

## Related issues

<!-- Recommended
Link related issues below. Insert the issue link or reference after the word "Closes" if merging this should automatically close it.

- Closes #123
- Relates #123
- Requires #123
- Superseds #123
-->
- Relates #37317